Intial work on transmitting responses and repeating unacknowledged messages.

+++ lib/protocol_parser/PcCommunicationsInterface.h
@@ -8,7 +8,7 @@ class PcCommunicationsInterface {
 
     public:
         virtual uint8_t GetReceivedBytes(uint8_t data[], uint8_t max_length) = 0;
-        virtual void TransmitData(uint8_t data[], uint8_t count) = 0;
+        virtual void TransmitData(uint8_t data[], uint16_t count) = 0;
 };
 
 #endif
